cleaned up expo router - fixed expo anchor warning (anchor within paragraph)
