+++ apps/expo/app/types.d.ts
@@ -1,0 +1,7 @@
+import { config } from '@my/config'
+
+export type Conf = typeof config
+
+declare module '@my/ui' {
+  interface TamaguiCustomConfig extends Conf {}
+}
