+++ apps/expo/app/user/[id].tsx
@@ -1,0 +1,24 @@
+import { UserDetailScreen } from 'app/features/user/detail-screen'
+import { Stack } from 'expo-router'
+import { useTheme } from '@my/ui'
+
+export default function Screen() {
+  const theme = useTheme()
+
+  return (
+    <>
+      <Stack.Screen
+        options={{
+          title: 'User',
+          presentation: 'modal',
+          animation: 'slide_from_right',
+          gestureEnabled: true,
+          gestureDirection: 'horizontal',
+          headerTintColor: theme.color.val,
+          headerStyle: { backgroundColor: theme.blue9.val },
+        }}
+      />
+      <UserDetailScreen />
+    </>
+  )
+}
