+++ apps/expo/tsconfig.json
@@ -5,7 +5,7 @@
     "app/**/*.tsx",
     ".expo/types/**/*.ts",
     "expo-env.d.ts"
-, "App.tsx"  ],
+  ],
   "compilerOptions": {
     "composite": true,
     "jsx": "react-jsx"
