+++ packages/app/provider/index.tsx
@@ -1,7 +1,6 @@
 import { useColorScheme } from 'react-native'
 import { CustomToast, TamaguiProvider, TamaguiProviderProps, ToastProvider, config } from '@my/ui'
 import { ToastViewport } from './ToastViewport'
-import { NavigationProvider } from './navigation'
 
 export function Provider({ children, ...rest }: Omit<TamaguiProviderProps, 'config'>) {
   const scheme = useColorScheme()
@@ -17,11 +16,9 @@ export function Provider({ children, ...rest }: Omit<TamaguiProviderProps, 'conf
           ]
         }
       >
-        <NavigationProvider>
-          {children}
-          <CustomToast />
-          <ToastViewport />
-        </NavigationProvider>
+        {children}
+        <CustomToast />
+        <ToastViewport />
       </ToastProvider>
     </TamaguiProvider>
   )
